requirements.txt
anyio==4.4.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
libtmux==0.37.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
PyYAML==6.0.2
rich==13.8.0
sniffio==1.3.1
